@@ -14,7 +14,8 @@ description: Flutter UI Kit @ Change
 version: 0.0.1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.10.0 <3.0.0"
+  flutter: ^1.22.0
 
 dependencies:
   flutter:
@@ -40,6 +41,7 @@ dev_dependencies:
   recase: ^2.0.1
   mockito: ^4.0.0
   package_info: ^0.4.0+3
+  junitreport: ^1.3.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
