@@ -2,22 +2,19 @@ import 'dart:math';
 
 import 'package:flutter/material.dart';
 
+import '../../color.dart';
+
 class GaugeIndicator extends StatelessWidget {
   final double width;
   final double bearishThreshold;
   final double bullishThreshold;
   final double score;
 
-  const GaugeIndicator({
-    @required this.width,
-    @required this.bearishThreshold,
-    @required this.bullishThreshold,
-    @required this.score
-  })
-      :assert(width > 0),
-       assert(bearishThreshold >= 0 && bearishThreshold < 1),
-       assert(bullishThreshold > 0 && bullishThreshold <= 1 && bullishThreshold > bearishThreshold),
-       assert(score >= 0 && score <= 1);
+  const GaugeIndicator({@required this.width, @required this.bearishThreshold, @required this.bullishThreshold, @required this.score})
+      : assert(width > 0),
+        assert(bearishThreshold >= 0 && bearishThreshold < 1),
+        assert(bullishThreshold > 0 && bullishThreshold <= 1 && bullishThreshold > bearishThreshold),
+        assert(score >= 0 && score <= 1);
 
   @override
   Widget build(BuildContext context) {
@@ -25,25 +22,23 @@ class GaugeIndicator extends StatelessWidget {
       width: width,
       height: width / 2,
       child: Stack(
-        children: <Widget>[
-          _gauge(context)
-        ],
-      ));
+        children: <Widget>[_gauge(context)],
+      ),
+    );
   }
 
   Widget _gauge(BuildContext context) {
     return Container(
-        width: width,
-        height: width / 2,
-        child: CustomPaint(
-          painter: _GaugePainter(width, width / 2, score, bearishThreshold, bullishThreshold),
-        )
+      width: width,
+      height: width / 2,
+      child: CustomPaint(
+        painter: _GaugePainter(width, width / 2, score, bearishThreshold, bullishThreshold),
+      ),
     );
   }
 }
 
 class _GaugePainter extends CustomPainter {
-
   final double width;
   final double height;
   final double score;
@@ -54,14 +49,12 @@ class _GaugePainter extends CustomPainter {
 
   @override
   void paint(Canvas canvas, Size size) {
-
     final paint = Paint()
       ..style = PaintingStyle.stroke
       ..strokeWidth = 2.5
       ..isAntiAlias = true;
 
-
-    final gaugeCenter = Offset(width/2, height);
+    final gaugeCenter = Offset(width / 2, height);
 
     final bgDiameter = width * 0.92;
     final bgRadius = bgDiameter / 2;
@@ -69,8 +62,11 @@ class _GaugePainter extends CustomPainter {
     final bgRect = Rect.fromCircle(center: gaugeCenter, radius: bgRadius);
     const Gradient gradient = LinearGradient(
       begin: Alignment.topCenter,
-        end: Alignment.bottomCenter,
-        colors: const [const Color(0x2172efdb), const Color(0x003edbb5)]
+      end: Alignment.bottomCenter,
+      colors: const [
+        const Color(0x2172efdb),
+        AppColor.ltGreenPrimary,
+      ],
     );
     final bgPaint = Paint()
       ..style = PaintingStyle.fill
@@ -81,13 +77,13 @@ class _GaugePainter extends CustomPainter {
     final rect = Rect.fromCircle(center: gaugeCenter, radius: width / 2);
     const delta = 0.05;
 
-    paint.color = const Color(0xffff6b6b);
+    paint.color = AppColor.red;
     canvas.drawArc(rect, pi, pi * bearishThreshold - delta, false, paint);
 
-    paint.color = const Color(0xff9da0a6);
+    paint.color = AppColor.semiGrey;
     canvas.drawArc(rect, pi + pi * bearishThreshold, pi * (bullishThreshold - bearishThreshold), false, paint);
 
-    paint.color = const Color(0xff4ce2a7);
+    paint.color = AppColor.brightGreen;
     canvas.drawArc(rect, 2 * pi - pi * (1 - bullishThreshold) + delta, pi * (1 - bullishThreshold) - delta, false, paint);
 
     final pointerCenter = gaugeCenter.translate(0, -7);
@@ -106,5 +102,4 @@ class _GaugePainter extends CustomPainter {
 
   @override
   bool shouldRepaint(CustomPainter oldDelegate) => false;
-
 }
