@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
-import 'copy_to_clipboard.dart';
+import 'package:flutter_ui_kit/widgets/text/copy_to_clipboard.dart';
+
+import '../../color.dart';
 import 'label_text.dart';
 
 class LabelValuePair extends StatelessWidget {
@@ -10,14 +12,7 @@ class LabelValuePair extends StatelessWidget {
   final bool copyToClipboardEnabled;
   final EdgeInsets padding;
 
-  const LabelValuePair({
-    @required this.labelText,
-    this.valueText,
-    this.copyToClipboardEnabled = false,
-    this.textAlign,
-    this.value,
-    this.padding
-  });
+  const LabelValuePair({@required this.labelText, this.valueText, this.copyToClipboardEnabled = false, this.textAlign, this.value, this.padding});
 
   @override
   Widget build(BuildContext context) {
@@ -25,9 +20,7 @@ class LabelValuePair extends StatelessWidget {
       children: [
         Expanded(
           child: Column(
-            crossAxisAlignment: _isCenterAligned(textAlign)
-                ? CrossAxisAlignment.center
-                : CrossAxisAlignment.start,
+            crossAxisAlignment: _isCenterAligned(textAlign) ? CrossAxisAlignment.center : CrossAxisAlignment.start,
             children: [
               LabelText(labelText),
               _buildBody(context, value, valueText, padding: padding),
@@ -46,11 +39,12 @@ class LabelValuePair extends StatelessWidget {
     return Padding(
       padding: padding ?? const EdgeInsets.only(bottom: 16.0),
       child: Row(
-        mainAxisAlignment: _isCenterAligned(textAlign)
-            ? MainAxisAlignment.center
-            : MainAxisAlignment.start,
+        mainAxisAlignment: _isCenterAligned(textAlign) ? MainAxisAlignment.center : MainAxisAlignment.start,
         children: [
-          Flexible(child: value != null ? value :  Text(data, style: const TextStyle(fontSize: 16 , fontFamily: 'DINNextLTPro', color: Colors.black, height: 1.5, fontWeight: FontWeight.w500))),
+          Flexible(
+              child: value != null
+                  ? value
+                  : Text(data, style: const TextStyle(fontSize: 16, fontFamily: 'DINNextLTPro', color: Colors.black, height: 1.5, fontWeight: FontWeight.w500))),
           _buildCopyToClipboardIcon(context, data),
         ],
       ),
@@ -61,10 +55,11 @@ class LabelValuePair extends StatelessWidget {
     if (!copyToClipboardEnabled) {
       return Container();
     }
-    return CopyToClipboard(value: value);
+    return CopyToClipboard(
+      value: value,
+      color: AppColor.green,
+    );
   }
-
 }
 
 enum ValueLabelTextAlign { center, left }
-
