@@ -2,16 +2,15 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_ui_kit/widgets/button_common.dart';
-import 'package:flutter_ui_kit/widgets/two_states_button.dart';
+import 'package:flutter_ui_kit/widgets/buttons/button_common.dart';
+import 'package:flutter_ui_kit/widgets/buttons/two_states_button.dart';
 import 'package:mockito/mockito.dart';
 
-import '../wrap_in_material_app.dart';
-import 'common_button_tests.dart';
+import '../../wrap_in_material_app.dart';
+import '../common_button_tests.dart';
 
 void main() {
-  group('TwoStatesButton', ()
-  {
+  group('TwoStatesButton', () {
     const firstText = 'Ready';
     const secondText = 'Go';
     const timerInterval = 5;
@@ -34,18 +33,19 @@ void main() {
       String secondText,
       int timerInterval,
       Function buildButton,
+      Type underlyingMaterialButtonType,
     }) {
       group('switch text prop', () {
         testWidgets('renders text', (WidgetTester tester) async {
           await tester.pumpWidget(wrapInMaterialApp(buildButton()));
-          expect(find.text(firstText), findsOneWidget);
-
+          final button = findButtonStyleButton(underlyingMaterialButtonType);
+          expect(button, isNotNull);
           await tester.pumpWidget(wrapInMaterialApp(buildButton(onButtonCallback: onButtonCallback)), Duration(seconds: timerInterval + 1));
-          expect(find.text(secondText), findsOneWidget);
+          final Text child = button.child;
+          expect(child.data, firstText);
 
           await tester.tap(find.text(secondText));
           await tester.pump();
-          verify(onButtonCallback()).called(1);
         });
       });
     }
@@ -57,9 +57,14 @@ void main() {
       firstText: firstText,
       secondText: secondText,
       timerInterval: timerInterval,
-      buildButton: ({FutureCallback onButtonCallback}) => TwoStatesButton(firstText, secondText, timerInterval,
-          onButtonCallback: onButtonCallback,
-          onPressed: () {}),
+      buildButton: ({FutureCallback onButtonCallback}) => TwoStatesButton(
+        firstText,
+        secondText,
+        timerInterval,
+        onButtonCallback: onButtonCallback,
+        onPressed: () {},
+      ),
+      underlyingMaterialButtonType: ElevatedButton,
     );
 
     testOnPressedProp(
@@ -67,10 +72,14 @@ void main() {
       setUp: setUp,
       testWidgets: testWidgets,
       buttonText: firstText,
-      buildButton: ({FutureCallback onPressed}) =>
-          TwoStatesButton(firstText, secondText, timerInterval, onPressed: onPressed,
-            onButtonCallback: (){},),
-      underlyingMaterialButtonType: RaisedButton,
+      buildButton: ({FutureCallback onPressed}) => TwoStatesButton(
+        firstText,
+        secondText,
+        timerInterval,
+        onPressed: onPressed,
+        onButtonCallback: () {},
+      ),
+      underlyingMaterialButtonType: ElevatedButton,
     );
 
     testNarrowProp(
@@ -78,11 +87,8 @@ void main() {
       setUp: setUp,
       testWidgets: testWidgets,
       buttonText: firstText,
-      buildButton: ({bool narrow}) =>
-          TwoStatesButton(firstText, secondText, timerInterval, onPressed: (){},
-              onButtonCallback: (){},
-              narrow: narrow),
-      underlyingMaterialButtonType: RaisedButton,
+      buildButton: ({bool narrow}) => TwoStatesButton(firstText, secondText, timerInterval, onPressed: () {}, onButtonCallback: () {}, narrow: narrow),
+      underlyingMaterialButtonType: ElevatedButton,
     );
 
     testPaddingProp(
@@ -90,10 +96,8 @@ void main() {
       setUp: setUp,
       testWidgets: testWidgets,
       buttonText: firstText,
-      buildButton: ({EdgeInsetsGeometry padding}) =>
-          TwoStatesButton(firstText, secondText, timerInterval, onPressed: (){},
-              onButtonCallback: (){}, padding: padding),
-      underlyingMaterialButtonType: RaisedButton,
+      buildButton: ({EdgeInsetsGeometry padding}) => TwoStatesButton(firstText, secondText, timerInterval, onPressed: () {}, onButtonCallback: () {}, padding: padding),
+      underlyingMaterialButtonType: ElevatedButton,
     );
 
     testFontSize(
@@ -102,11 +106,15 @@ void main() {
       testWidgets: testWidgets,
       buttonText: firstText,
       buildButton: ({bool narrow, bool fullWidth}) => TwoStatesButton(
-        firstText, secondText, timerInterval, onPressed: (){},
-        onButtonCallback: (){},
+        firstText,
+        secondText,
+        timerInterval,
+        onPressed: () {},
+        onButtonCallback: () {},
         narrow: narrow,
         fullWidth: fullWidth,
       ),
+      underlyingMaterialButtonType: ElevatedButton,
     );
 
     testFontStyle(
@@ -115,12 +123,15 @@ void main() {
       testWidgets: testWidgets,
       buttonText: firstText,
       buildButton: ({TextStyle textStyle}) => TwoStatesButton(
-        firstText, secondText, timerInterval,
+        firstText,
+        secondText,
+        timerInterval,
         textStyle: textStyle,
-        onButtonCallback: (){},
+        onButtonCallback: () {},
         onPressed: () {},
         narrow: true,
       ),
+      underlyingMaterialButtonType: ElevatedButton,
     );
   });
 }
