@@ -1,5 +1,7 @@
 import 'dart:math';
 
+import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:flutter_ui_kit/color.dart';
 import 'package:flutter_ui_kit/story_book/expandable_story.dart';
 import 'package:flutter_ui_kit/story_book/prop_updater/bool_prop_updater.dart';
@@ -7,12 +9,11 @@ import 'package:flutter_ui_kit/story_book/prop_updater/int_prop_updater.dart';
 import 'package:flutter_ui_kit/story_book/prop_updater/list_prop_updater.dart';
 import 'package:flutter_ui_kit/story_book/prop_updater/string_prop_updater.dart';
 import 'package:flutter_ui_kit/story_book/props_explorer.dart';
-import 'package:flutter_ui_kit/widgets/filled_button.dart';
-import 'package:flutter_ui_kit/widgets/outlined_button.dart';
-import 'package:flutter/material.dart';
+import 'package:flutter_ui_kit/widgets/buttons/outlined_button.dart';
+import 'package:flutter_ui_kit/widgets/buttons/text_button.dart';
+import 'package:flutter_ui_kit/widgets/buttons/filled_button.dart';
 import 'package:flutter_ui_kit/widgets/switcher_button.dart';
-import 'package:flutter_ui_kit/widgets/text_button.dart';
-import 'package:flutter_ui_kit/widgets/two_states_button.dart';
+import 'package:flutter_ui_kit/widgets/buttons/two_states_button.dart';
 
 class Buttons extends StatelessWidget {
   @override
@@ -25,7 +26,7 @@ class Buttons extends StatelessWidget {
           _outlinedButtonStory(),
           _textButtonStory(),
           _switcherButtonStory(),
-          _twoStatesButtonStory()
+          _twoStatesButtonStory(),
         ],
       ),
     );
@@ -95,7 +96,13 @@ class Buttons extends StatelessWidget {
             };
           }
 
-          const colors = [Colors.red, Colors.blue, Colors.purple, Colors.teal, Colors.brown];
+          const colors = [
+            Colors.red,
+            Colors.blue,
+            Colors.purple,
+            Colors.teal,
+            Colors.brown,
+          ];
           final randIndex = Random().nextInt(5);
 
           return FilledButton(
@@ -175,10 +182,9 @@ class Buttons extends StatelessWidget {
           }
 
           return Container(
-            width: double.infinity,
             padding: const EdgeInsets.symmetric(vertical: 10),
-            color: props['alt'] ? AppColor.green : Colors.white,
-            child: OutlinedButton(
+            color: props['alt'] ? AppColor.green : AppColor.ltDeepWhite,
+            child: BorderedButton(
               props['text'],
               onPressed: onPressed,
               fullWidth: props['fullWidth'],
@@ -245,8 +251,8 @@ class Buttons extends StatelessWidget {
 
           return Container(
             width: double.infinity,
-            color: props['alt'] ? AppColor.green : Colors.white,
-            child: TextButton(
+            color: props['alt'] ? AppColor.green : AppColor.ltDeepWhite,
+            child: PlainButton(
               props['text'],
               onPressed: onPressed,
               alt: props['alt'],
