@@ -1,9 +1,10 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_ui_kit/widgets/filled_button.dart';
+import 'package:flutter/material.dart' as mat;
+import 'package:flutter/widgets.dart';
+import 'package:flutter_ui_kit/widgets/buttons/text_button.dart';
+import 'package:flutter_ui_kit/widgets/buttons/filled_button.dart';
 import 'package:flutter_ui_kit/widgets/popover_modal.dart';
 import 'package:flutter_ui_kit/widgets/text/heading.dart';
 import 'package:flutter_ui_kit/widgets/text/heading_type.dart';
-import 'package:flutter_ui_kit/widgets/text_button.dart';
 
 class Popover extends StatelessWidget {
   final PopoverModal popover = PopoverModal();
@@ -18,7 +19,7 @@ class Popover extends StatelessWidget {
           children: [
             _popoverExampleOne(context),
             _popoverExampleTwo(context),
-            _popoverExampleThree(context)
+            _popoverExampleThree(context),
           ],
         ),
       ),
@@ -80,8 +81,7 @@ class Popover extends StatelessWidget {
           padding: const EdgeInsets.all(30.0),
           child: Column(
             children: const <Widget>[
-              const Heading('Example Header',
-                  headingType: HeadingType.HEADING_2),
+              const Heading('Example Header', headingType: HeadingType.HEADING_2),
               Padding(
                 padding: const EdgeInsets.all(8.0),
                 child: const Text(
@@ -102,8 +102,7 @@ class Popover extends StatelessWidget {
           child: Column(
             mainAxisSize: MainAxisSize.min,
             children: const <Widget>[
-              const Heading('Example Header',
-                  headingType: HeadingType.HEADING_2),
+              const Heading('Example Header', headingType: HeadingType.HEADING_2),
               Padding(
                 padding: const EdgeInsets.all(8.0),
                 child: const Text(
@@ -120,13 +119,11 @@ class Popover extends StatelessWidget {
     final result = popover.showPopover<String>(
         context,
         Padding(
-          padding: const EdgeInsets.only(
-              left: 30.0, top: 30.0, bottom: 5.0, right: 30.0),
+          padding: const EdgeInsets.only(left: 30.0, top: 30.0, bottom: 5.0, right: 30.0),
           child: Column(
             mainAxisSize: MainAxisSize.min,
             children: <Widget>[
-              const Heading('Example Header',
-                  headingType: HeadingType.HEADING_2),
+              const Heading('Example Header', headingType: HeadingType.HEADING_2),
               const Padding(
                 padding: const EdgeInsets.all(12.0),
                 child: const Text(
@@ -146,7 +143,7 @@ class Popover extends StatelessWidget {
               ),
               Padding(
                 padding: const EdgeInsets.all(8.0),
-                child: TextButton('Cancel', onPressed: () {
+                child: PlainButton('Cancel', onPressed: () {
                   Navigator.pop(context, 'Cancel button');
                 }),
               )
@@ -156,14 +153,14 @@ class Popover extends StatelessWidget {
 
     result.then((value) {
       value ??= 'Nothing';
-      showDialog<void>(
+      mat.showDialog<void>(
           context: context,
           builder: (BuildContext context) {
-            return AlertDialog(
+            return mat.AlertDialog(
               title: const Text('You clicked'),
               content: Text(value),
               actions: <Widget>[
-                FlatButton(
+                mat.FlatButton(
                     child: const Text('Thats nice'),
                     onPressed: () {
                       Navigator.pop(context);
