@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'colors.dart';
+import '../color.dart';
 
 class MainAppBar extends StatelessWidget implements PreferredSizeWidget {
   final Widget title;
@@ -10,7 +10,14 @@ class MainAppBar extends StatelessWidget implements PreferredSizeWidget {
   final Color backgroundColor;
   final List<Widget> actions;
 
-  const MainAppBar({this.title, this.leading, this.centerTitle = false, this.implyLeading = true, this.backgroundColor, this.actions = const <Widget>[]});
+  const MainAppBar({
+    this.title,
+    this.leading,
+    this.centerTitle = false,
+    this.implyLeading = true,
+    this.backgroundColor,
+    this.actions = const <Widget>[],
+  });
 
   @override
   Size get preferredSize => const Size.fromHeight(kToolbarHeight);
@@ -25,8 +32,7 @@ class MainAppBar extends StatelessWidget implements PreferredSizeWidget {
       backgroundColor: backgroundColor ?? Colors.white,
       elevation: 0.0,
       actions: actions,
-      iconTheme:
-          Theme.of(context).iconTheme.copyWith(color: AppColors.secondary),
+      iconTheme: Theme.of(context).iconTheme.copyWith(color: AppColor.ltGreenPrimary),
     );
   }
 }
