@@ -4,8 +4,12 @@ import 'package:flutter_ui_kit/ui/text/subhead.dart';
 
 void main() {
   testWidgets('Subhead shows text', (tester) async {
-    await tester.pumpWidget(const Directionality(
-        textDirection: TextDirection.ltr, child: const Subhead('Test')));
+    await tester.pumpWidget(
+      const Directionality(
+        textDirection: TextDirection.ltr,
+        child: const Subhead('Test'),
+      ),
+    );
     expect(find.text('Test'), findsOneWidget);
   });
 }
