@@ -5,14 +5,12 @@ import 'package:flutter_test/flutter_test.dart';
 final _log = <MethodCall>[];
 final _responses = <MethodCallStubbing, dynamic>{};
 
-void setUpTestMethodChannel(String methodChannel,
-    [MethodCodec codec = const StandardMethodCodec()]) {
+void setUpTestMethodChannel(String methodChannel, [MethodCodec codec = const StandardMethodCodec()]) {
   debugPrint('Setting up test method channel $methodChannel');
   final channel = MethodChannel(methodChannel, codec);
   channel.setMockMethodCallHandler((methodCall) async {
     _log.add(methodCall);
-    final matchingStubbing = _responses.keys
-        .firstWhere((s) => s.matches(methodCall), orElse: () => null);
+    final matchingStubbing = _responses.keys.firstWhere((s) => s.matches(methodCall), orElse: () => null);
     if (matchingStubbing != null) {
       return _responses[matchingStubbing];
     }
@@ -35,8 +33,7 @@ class MethodCallStubbing {
   }
 
   bool matches(MethodCall methodCall) {
-    return nameMatcher.matches(methodCall.method, null) &&
-        argMatcher.matches(methodCall.arguments, null);
+    return nameMatcher.matches(methodCall.method, null) && argMatcher.matches(methodCall.arguments, null);
   }
 }
 
@@ -44,13 +41,13 @@ MethodCallStubbing whenMethodCall(Matcher nameMatcher, Matcher argMatcher) {
   return MethodCallStubbing(nameMatcher, argMatcher);
 }
 
-void expectMethodCall(String name, {Map<String, Object> arguments}) {
+void expectMethodCall(String name, {@required dynamic args}) {
   final methodCall = _log.firstWhere((call) => call.method == name, orElse: () {
     fail('No call found for $name, all calls $_log');
   });
 
   expect(
     methodCall,
-    isMethodCall(name, arguments: arguments),
+    isMethodCall(name, arguments: args),
   );
 }
