@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_ui_kit/ui/text/copy_to_clipboard.dart';
+import 'package:flutter_ui_kit/widgets/text/copy_to_clipboard.dart';
 import 'package:mockito/mockito.dart';
 
 import '../test_method_channel.dart';
@@ -14,9 +14,7 @@ class MockVoidFunction extends Mock implements Function {
 void main() {
   group('Copy To Clipboard', () {
     testWidgets('renders icon', (tester) async {
-      await tester.pumpWidget(const Directionality(
-          textDirection: TextDirection.ltr,
-          child: const CopyToClipboard(value: '123456')));
+      await tester.pumpWidget(const Directionality(textDirection: TextDirection.ltr, child: const CopyToClipboard(value: '123456')));
       expect(find.byType(Icon), findsOneWidget);
     });
 
@@ -27,7 +25,7 @@ void main() {
         home: Scaffold(body: CopyToClipboard(value: '123456', onTapCallback: onTapCallback.call)),
       ));
       await tester.tap(find.byType(Icon));
-      expectMethodCall('Clipboard.setData', arguments: <String, dynamic>{
+      expectMethodCall('Clipboard.setData', args: <String, dynamic>{
         'text': '123456',
       });
       verify(onTapCallback.call());
